Basic /convert functionality
The API /convert route should, in theory, work now. Testing (and inevitable bugfixes) to come later.
