+++ deprecated/schemas/0.4.0_repository.schema
