+++ deprecated/schemas/data_type_tags.json
