+++ deprecated/schemas/generate_schemas.py
