+++ mdf_refinery/record.schema
@@ -1,0 +1,1 @@
+{}
